Store price as int rather than float

+++ backend/economicalc/objects/item.py
@@ -1,5 +1,5 @@
 class Item:
-    def __init__(self, item_name: str, price: float) -> None:
+    def __init__(self, item_name: str, price: int) -> None:
         self.item_name = item_name
         self.price = price
